fullConfig/requirements.txt
paho-mqtt
psycopg2-binary
opensearch-py